bug fixes and utils code reorg
