+++ config.js
@@ -1,10 +1,10 @@
 module.exports = {
     'mainnet': {
-        metadata_url: "https://nft.f1deltatime.com",
+        assets_url: "https://nft.f1deltatime.com",
         external_url: "https://f1deltatime.com/nft-preview?id={id}"
     },
     'rinkeby': {
-        metadata_url: "https://nft-test.f1deltatime.com",
+        assets_url: "https://nft-test.f1deltatime.com",
         external_url: "https://f1deltatime.com/nft-preview?id={id}"
     }
 }
