+++ package.json
@@ -14,6 +14,7 @@
   "license": "MIT",
   "dependencies": {
     "@animoca/blockchain-inventory_metadata": "^0.0.1",
+    "big-integer": "^1.6.48",
     "bits.js": "^0.1.1",
     "commander": "^5.1.0"
   },
