+++ src/mappings/Season2019/TokenTypes/Gear/Gears.js
@@ -1,5 +1,5 @@
 const Season = require('../../Season');
-const { createCollectionId } = require('../../../../metadata/utils/collection');
+const { createCollectionId } = require('../../../../utils/ids');
 
 const Gloves = {
     subType: 'Gloves',
