+++ src/mappings/Season2019/TokenTypes/Tyres/Tyres.js
@@ -1,5 +1,5 @@
 const Season = require('../../Season');
-const { createCollectionId } = require('../../../../metadata/utils/collection');
+const { createCollectionId } = require('../../../../utils/ids');
 
 const SoftTyres = {
     subType: 'Soft Tyres',
