--- src/metadata/index.js
@@ -1,7 +1,0 @@
-const { createTokenId, getCoreMetadata, getFullMetadata } = require('./metadata');
-
-module.exports = {
-    createTokenId,
-    getCoreMetadata,
-    getFullMetadata,
-};
