--- src/metadata/utils/index.js
@@ -1,10 +1,0 @@
-const { getTokenImage } = require('./image');
-const { createCollectionId } = require('./collection');
-const { validateCommonMetadata, validateSeasonMetadata } = require('./validation');
-
-module.exports = {
-    createCollectionId,
-    getTokenImage,
-    validateCommonMetadata,
-    validateSeasonMetadata,
-};
